pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "aspire-bilby"
authors = [
    {name = "Michael J. Williams", email = "michaeljw1@googlemail.com"},
]
description = ""
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "bilby>=2.7.0",
    "aspire-inference>=0.1.0a16",
    "numpy",
    "pandas",
]

dynamic = ["version"]

[project.optional-dependencies]
test = [
    "pytest",
]
torch = [
    "aspire-inference[torch]",
]
jax = [
    "aspire-inference[jax]",
]
minipcn = [
    "minipcn[array-api]>=0.2.0a3",
]

[tool.setuptools_scm]

[tool.black]
line-length = 79

[project.entry-points."bilby.samplers"]
aspire = "aspire_bilby.plugin:Aspire"
